feat(core/dfn-panel): run in exported documents
